Work on performance

+++ src/main/java/pl/sokolak/teamtally/frontend/user_section/ranking/dto/TeamWithPlace.java
@@ -1,12 +1,11 @@
 package pl.sokolak.teamtally.frontend.user_section.ranking.dto;
 
-import java.util.HashSet;
-import java.util.Set;
+import java.math.BigDecimal;
 
 public record TeamWithPlace(String name,
                             String icon,
                             String color,
-                            int points,
+                            BigDecimal points,
                             int place) {
 
     public TeamWithPlace(TeamWithPoints team, int place) {
